refactor: fix eslint errors
