+++ src/data/serialize-utils.ts
@@ -5,12 +5,13 @@
  * @param obj1
  * @param obj2
  */
+//eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const combineObjects = (obj1: any, obj2: any) => {
 	// Create a new object with the properties of obj1
 	const result = { ...obj1 };
 
 	for (const key in obj1) {
-		if (!obj2.hasOwnProperty(key)) {
+		if (!(key in obj2)) {
 			// Remove the key if it doesn't exist in the second object
 			delete result[key];
 		}
