+++ src/obsidian/dashboards-settings-tab.ts
@@ -1,4 +1,4 @@
-import { PluginSettingTab, App, Setting } from "obsidian";
+import { PluginSettingTab, App, Setting, setIcon } from "obsidian";
 import DashboardsPlugin from "src/main";
 
 export default class DashboadsSettingsTab extends PluginSettingTab {
@@ -22,10 +22,12 @@ export default class DashboadsSettingsTab extends PluginSettingTab {
 		//Attachments folder
 		const attachmentsFolderDesc = new DocumentFragment();
 		attachmentsFolderDesc.createSpan({}, (span) => {
-			span.innerHTML = `Create dashboards in the attachments folder defined in the Obsidian settings.<br><br>This can be changed in <span style="color: var(--text-accent);">Files & Links -> Default location for new attachments</span><br><br>Otherwise, the folder location below will be used.`;
+			span.innerHTML =
+				// eslint-disable-next-line quotes
+				'Create dashboards in the attachments folder defined in the Obsidian settings.<br><br>This can be changed in <span style="color: var(--text-accent);">Files & Links -> Default location for new attachments</span><br><br>Otherwise, the folder location below will be used.';
 		});
 
-		new Setting(containerEl)
+		const setting = new Setting(containerEl)
 			.setName("Create in attachments folder")
 			.setDesc(attachmentsFolderDesc)
 			.addToggle((cb) => {
@@ -38,11 +40,16 @@ export default class DashboadsSettingsTab extends PluginSettingTab {
 					this.display();
 				});
 			});
+		const settingContainerEl = setting.controlEl;
+		const div = document.createElement("div");
+		setIcon(div, "lock");
+		settingContainerEl.insertBefore(div, settingContainerEl.firstChild);
 
 		//Folder location
 		const defaultLocationDesc = new DocumentFragment();
 		defaultLocationDesc.createSpan({}, (span) => {
-			span.innerHTML = `Where newly created dashboards are placed. Default location is the vault root folder, if not specified.`;
+			span.innerHTML =
+				"Where newly created dashboards are placed. Default location is the vault root folder, if not specified.";
 		});
 
 		if (this.plugin.settings.createInObsidianAttachmentFolder === false) {
